Fix copy game code to clipboard button
Firefox does not define the ClipboardItem object by default.

Users can toggle asyncClipboard.clipboardItem in about:config to enable.

It's not realistic or recommended to ask users to do this.

Thankfully it's not needed if we call clipboard.writeText instead.

Copy button should now work in Firefox and Chromium based browsers.

Safari appears to still have issues but I'm not sure why yet.