Wire up global key listener in layout to store
This allows us to define only one global key listener for the whole app.

It can decide which events are propagated to the children.

Children can then respond to the event by subscribing to the store.

That way they don't need to add their own individual event listeners.

This fixes the bug where closing the Info Center also closes the player.