Fix upload button not reuploading discarded file
Uploaded files are kept in the input's files list for submission.

File parsing is wired up to the input's onChange event.

Because the file is already in the list it does not change.

Because the files list does not change the event does not fire.

As a result, the file cannot be reuploaded once discarded.

As we are not submitting a form, we can discard the file after parsing.

This allows it to be reuploaded if necessary.